[accumulators] Reshuffle and move code from wrapper_set.hpp to .cpp files
New .cpp files:

    /accumulators/src/wrapper_set.cpp
    /accumulators/src/wrapper_set_hdf5.cpp
